Revert "Bug 2042256 - Adjust fuzz for Skia m149. r=gfx-reviewers,media-playback-reviewers,nical,aosmond" for causing xpc failures at test_imgtools.js
This reverts commit 9f778821407053011a05ddc785f678fd27ace0bf.

Revert "Bug 2042256 - Update mozbuilds for Skia m149. r=gfx-reviewers,nical"

This reverts commit 674dfdde151ddaeae487eb5735d4a320631584b7.

Revert "Bug 2042256 - Fix Moz2D for Skia m149. r=gfx-reviewers,nical"

This reverts commit c9d727ae880e00e97e291f346db252d7856c2193.

Revert "Bug 2042256 - Update Skia source files for Skia m149. r=gfx-reviewers,nical"

This reverts commit dc99955deb8172c40e307661a01b2407bd95ff23.

+++ gfx/skia/skia/src/core/SkVertices.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2017 Google Inc.
  *
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
@@ -153,7 +153,7 @@ void SkVertices::Builder::init(const Desc& desc) {
 sk_sp<SkVertices> SkVertices::Builder::detach() {
     if (fVertices) {
         fVertices->fBounds = SkRect::BoundsOrEmpty({fVertices->fPositions,
-                                                    (size_t)fVertices->fVertexCount});
+                                                    fVertices->fVertexCount});
         if (fVertices->fMode == kTriangleFan_VertexMode) {
             if (fIntermediateFanIndices) {
                 SkASSERT(fVertices->fIndexCount);
