+++ gfx/skia/skia/src/utils/SkPolyUtils.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2017 Google Inc.
  *
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
@@ -1641,7 +1641,7 @@ bool SkTriangulateSimplePolygon(const SkPoint* polygonVerts, uint16_t* indexMap,
 
     // get bounds
     SkRect bounds;
-    if (!bounds.setBoundsCheck({polygonVerts, (size_t)polygonSize})) {
+    if (!bounds.setBoundsCheck({polygonVerts, polygonSize})) {
         return false;
     }
     // get winding direction
